+++ inst/doc/quick_start.R
@@ -7,23 +7,25 @@ library(foghorn)
 summary_cran_results(email = "francois.michonneau@gmail.com")
 
 ## ----tibble--------------------------------------------------------------
-## Summary as a tibble
-check_cran_results(email = "francois.michonneau@gmail.com")
+## Results as a tibble
+cran_results(email = "francois.michonneau@gmail.com")
 
 ## ----package-name--------------------------------------------------------
 ## either by themselves
 summary_cran_results(pkg = c("ggplot2", "dplyr"))
-check_cran_results(pkg = c("ggplot2", "dplyr"))
+cran_results(pkg = c("ggplot2", "dplyr"))
 
 ## or by combining them with email addresses
 summary_cran_results(email = "francois.michonneau@gmail.com",
                      pkg = c("mregions", "ridigbio"))
-check_cran_results(email = "francois.michonneau@gmail.com",
+cran_results(email = "francois.michonneau@gmail.com",
                    pkg = c("mregions", "ridigbio"))
 
-## ----show-cran-results---------------------------------------------------
-show_cran_results(pkg = "tidyr")
+## ----cran-details--------------------------------------------------------
+tidyr_details <- cran_details(pkg = "tidyr")
+tidyr_details
+summary(tidyr_details)
 
 ## ----show-cran-db--------------------------------------------------------
-check_cran_results(pkg = "nlme", src = "crandb")
+cran_results(pkg = "nlme", src = "crandb")